refactor: move order_graph_manager into client/order directory
